Relocate `CreatePluginCommand` `ListPluginsCommand` `PluginInfoCommand` from `org.grails.cli.commands` to `org.grails.cli.command.plugin`
