+++ grace-shell/src/main/groovy/org/grails/cli/command/plugin/PluginInfoCommand.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands
+package org.grails.cli.command.plugin
 
 import groovy.xml.XmlSlurper
 import groovy.xml.slurpersupport.GPathResult
@@ -21,6 +21,11 @@ import groovy.xml.slurpersupport.GPathResult
 import grails.build.logging.GrailsConsole
 
 import org.grails.build.parsing.CommandLine
+import org.grails.cli.commands.Command
+import org.grails.cli.commands.CommandDescription
+import org.grails.cli.commands.ExecutionContext
+import org.grails.cli.commands.GlobalCommand
+import org.grails.cli.commands.ProjectCommand
 
 /**
  * Get the information about the given plugin
