+++ grace-shell/src/main/resources/META-INF/services/org.grails.cli.commands.Command
@@ -3,14 +3,12 @@ org.grails.cli.command.app.CreateAppCommand
 org.grails.cli.command.destroy.DestroyCommand
 org.grails.cli.command.generate.GenerateCommand
 org.grails.cli.command.gradle.GradleCommand
+org.grails.cli.command.plugin.CreatePluginCommand
+org.grails.cli.command.plugin.ListPluginsCommand
+org.grails.cli.command.plugin.PluginInfoCommand
 org.grails.cli.command.run.RunCommand
-org.grails.cli.commands.ApplicationTemplateCommand
-org.grails.cli.commands.CreateAppCommand
-org.grails.cli.commands.CreatePluginCommand
 org.grails.cli.commands.CreateProfileCommand
 org.grails.cli.commands.OpenCommand
 org.grails.cli.commands.HelpCommand
-org.grails.cli.commands.ListPluginsCommand
 org.grails.cli.commands.ListProfilesCommand
-org.grails.cli.commands.PluginInfoCommand
 org.grails.cli.commands.ProfileInfoCommand
